requirements.txt
Pillow
pycryptodomex